% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_MCED_screen_outputs}
\alias{calc_MCED_screen_outputs}
\title{Calculate MCED outcomes}
\usage{
calc_MCED_screen_outputs(markov_trace)
}
\arguments{
\item{markov_trace}{Matrix of MCED patient state occupancy}
}
\value{
total number tested, total FPs, and total MCED screening costs
}
\description{
Calculates the number tested, FPs, and costs for MCED screening tests
}
\examples{
calc_MCED_screen_outputs()
}
