% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_no_diag_health_outcomes}
\alias{calc_no_diag_health_outcomes}
\title{Calculate life years and QALYS gained for never-cancer patient}
\usage{
calc_no_diag_health_outcomes(total_diagnosed)
}
\arguments{
\item{total_diagnosed}{Matrix of patient state occupancy}
}
\value{
dataframe with undiscounted LYs, discounted LYs, undiscounted QALYs, discounted QALYs
}
\description{
Calculate life years and QALYS by patients who never developed cancer
}
\examples{
calc_no_diag_health_outcomes()
}
