% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_post_diag_res_MCED}
\alias{calc_post_diag_res_MCED}
\title{Calculates LYs, QALYs, deaths, and costs after MCED diagnosis}
\usage{
calc_post_diag_res_MCED(
  iCycle,
  iCancer,
  iStage,
  iLeadTime,
  iPts,
  res_LYs,
  res_QALYs,
  res_costs,
  res_societal_costs,
  res_deaths,
  calc_year,
  maxScreening,
  didnt_shift
)
}
\arguments{
\item{iCycle}{x}

\item{iCancer}{x}

\item{iStage}{x}

\item{iLeadTime}{x}

\item{iPts}{x}

\item{res_LYs}{x}

\item{res_QALYs}{x}

\item{res_costs}{x}

\item{res_societal_costs}{x}

\item{res_deaths}{x}

\item{calc_year}{x}

\item{maxScreening}{x}
}
\value{
assigns to existing variables?
}
\description{
Calculates LYs, QALYs, deaths, and costs after MCED diagnosis
}
\examples{
calc_post_diag_res_MCED()
}
