% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_post_diag_res_MCED_TS}
\alias{calc_post_diag_res_MCED_TS}
\title{Calculates LYs, QALYs, deaths, and costs after MCED diagnosis for time shifted individuals}
\usage{
calc_post_diag_res_MCED_TS(
  iCycle,
  iCancer,
  iStage,
  iLeadTime,
  iPts,
  res_LYs_TS,
  res_QALYs_TS,
  res_costs_TS,
  calc_year,
  maxScreening,
  didnt_shift
)
}
\arguments{
\item{iCycle}{x}

\item{iCancer}{x}

\item{iStage}{x}

\item{iLeadTime}{x}

\item{iPts}{x}

\item{res_LYs_TS}{x}

\item{res_QALYs_TS}{x}

\item{res_costs_TS}{x}

\item{calc_year}{x}

\item{maxScreening}{x}

\item{cost_growth_rate}{x}
}
\value{
assigns to existing variables?
}
\description{
Calculates LYs, QALYs, deaths, and costs after MCED diagnosis for time shifted individuals
}
\examples{
calc_post_diag_res_MCED_TS()
}
