% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_post_diag_res_overdiag}
\alias{calc_post_diag_res_overdiag}
\title{Calculates QALYS and costs associated with overdiagnosis
Calculates QALYS and costs associated with overdiagnosis}
\usage{
calc_post_diag_res_overdiag(
  iCycle,
  iCancer,
  iStage,
  mean_time_shift,
  iLeadTime,
  iPts,
  res_costs,
  res_QALYs,
  res_screen
)
}
\arguments{
\item{iCycle}{x}

\item{iCancer}{x}

\item{iStage}{x}

\item{mean_time_shift}{x}

\item{iLeadTime}{x}

\item{iPts}{x}

\item{res_costs}{x}

\item{res_QALYs}{x}

\item{res_screen}{x}
}
\value{
assigns to existing variables?
}
\description{
Calculates QALYS and costs associated with overdiagnosis
Calculates QALYS and costs associated with overdiagnosis
}
\examples{
calc_post_diag_res_overdiag()
}
