% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_post_diag_util}
\alias{calc_post_diag_util}
\title{Calculates discounted QALYS gained after diagnosis}
\usage{
calc_post_diag_util(
  surv,
  diag_age,
  perc_iatrogenic,
  unit,
  disc_rate,
  flag_mced,
  modifier
)
}
\arguments{
\item{surv}{Years lived after diagnosis}

\item{diag_age}{age at diagnosis}

\item{perc_iatrogenic}{Pre-specified percentage who receive iatrogenic harm}

\item{unit}{???/}

\item{disc_rate}{Pre-specified health discount rate}

\item{flag_mced}{Whether this is for MCED or SoC scenario}
}
\value{
returns with discounted QALYs
}
\description{
Calculate discounted QALYS gained by patients after cancer diagnosis
}
\examples{
calc_post_diag_util()
}
