% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_res_post_diag}
\alias{calc_res_post_diag}
\title{Calculate discounted life years gained after diagnosis}
\usage{
calc_res_post_diag(surv, diag_age, perc_iatrogenic, unit, disc_rate, flag_mced)
}
\arguments{
\item{surv}{Years lived after diagnosis}

\item{diag_age}{age at diagnosis}

\item{perc_iatrogenic}{Pre-specified percentage who receie iatrogenic harm}

\item{unit}{???/}

\item{disc_rate}{Pre-specified health discount rate}

\item{flag_mced}{Whether this is for MCED or SoC scenario}
}
\value{
returns with discounted LYs
}
\description{
Calculate discounted life years gained by patients after cancer diagnosis
}
\examples{
calc_res_post_diag()
}
