% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_res_pre_diag}
\alias{calc_res_pre_diag}
\title{Calculate life years and QALYS gained pre-diagnosis}
\usage{
calc_res_pre_diag(markov_trace)
}
\arguments{
\item{markov_trace}{Matrix of patient state occupancy}
}
\value{
dataframe with sum_stages, undiscounted LYs, discounted LYs, undiscounted QALYs, discounted QALYs
}
\description{
Calculate life years and QALYS gained pre-diagnosis
}
\examples{
calc_res_pre_diag()
}
