% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{calc_soc_screen_outputs}
\alias{calc_soc_screen_outputs}
\title{Calculate Standard of Care outcomes}
\usage{
calc_soc_screen_outputs(markov_trace)
}
\arguments{
\item{markov_trace}{Matrix of SoC patient state occupancy}
}
\value{
total number tested, total FPs, and total SoC screening costs
}
\description{
Calculates the number tested, FPs, and costs for SoC screening tests
}
\examples{
calc_soc_screen_outputs()
}
