% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interception_model_functions.R
\name{calc_stage_shift_matrix}
\alias{calc_stage_shift_matrix}
\title{Calculates stage shift matrix}
\usage{
calc_stage_shift_matrix(
  param_dwell_times,
  param_screen_interval,
  param_sensitivity,
  param_incidence
)
}
\arguments{
\item{param_dwell_times}{dataframe of dwell times in years by cancer and by stage}

\item{param_screen_interval}{interval between MCED tests, e.g years}

\item{param_sensitivity}{dataframe of MCED test sensitivity by cancer and by stage}

\item{param_incidence}{dataframe of annual cancer incidence rates}
}
\value{
returns stage shift matrix for all cancers in 4x4 format
}
\description{
Executes InterceptModel calculations to get stage shift matrix
}
\examples{
calc_stage_shift_matrix()
}
