% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{discrete_discount}
\alias{discrete_discount}
\title{Calculates discretely discounted area}
\usage{
discrete_discount(start_value, start_time, end_time, disc_rate)
}
\arguments{
\item{start_value}{x}

\item{start_time}{x}

\item{end_time}{x}

\item{disc_rate}{x}
}
\value{
discounted area
}
\description{
Calculate discretely discounted area
}
\examples{
discrete_discount()
}
