% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interception_model_functions.R
\name{effective_sens}
\alias{effective_sens}
\title{Calculates}
\usage{
effective_sens(cumulative_sens, dwell_detect_rate)
}
\arguments{
\item{cumulative_sens}{x}

\item{dwell_detect_rate}{x}
}
\value{
returns list(intercept = detect, clinical = live, parrive = arrive, plive = live, pdetect = detect, pmiss = miss)
}
\description{
Calculates ... based on the effective sensitivity, i.e taking into account
}
\examples{
effective_sens()
}
