% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_death60_summary_table.R
\name{f_death60_summary_table}
\alias{f_death60_summary_table}
\title{Summary table across groups for probability of survival past 60}
\usage{
f_death60_summary_table(list)
}
\arguments{
\item{list}{List of outcome dataframes}
}
\value{
Summary table
}
\description{
Summary table across groups for probability of survival past 60
}
\examples{
f_death60_summary_table()
}
