% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_hazard_split.R
\name{f_derive_netsurv_cfDNAstatus}
\alias{f_derive_netsurv_cfDNAstatus}
\title{Derive differential survival by cfDNA status}
\usage{
f_derive_netsurv_cfDNAstatus(net_survival, hazard_ratio, cfDNA_detection_rate)
}
\arguments{
\item{net_survival}{Survival probability}

\item{hazard_ratio}{Hazard ratio between cfDNA negative and positive status}

\item{cfDNA_detection_rate}{Proportion of cfDNA postivie who are detected}
}
\value{
Estimate of cfDNA negative survival
}
\description{
Calculates differential survival based on hazard ratio between cfDNA statuses
}
\examples{
f_derive_netsurv_cfDNAstatus()
}
