% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_LE_errors.R
\name{f_LE_errors}
\alias{f_LE_errors}
\title{Catch errors}
\usage{
f_LE_errors(list)
}
\arguments{
\item{list}{List of outcome dataframes}
}
\value{
Summary table of errors
}
\description{
Catch errors
}
\examples{
f_LE_errors()
}
