% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_LE_summary_table.R
\name{f_LE_summary_table}
\alias{f_LE_summary_table}
\title{Summary table across groups}
\usage{
f_LE_summary_table(list)
}
\arguments{
\item{list}{List of outcome dataframes}
}
\value{
Summary table
}
\description{
Summary table across groups
}
\examples{
f_LE_summary_table()
}
