% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_life_exp_from_5year_netsurv.R
\name{f_life_exp_from_5year_netsurv}
\alias{f_life_exp_from_5year_netsurv}
\title{Modal life expectancy using M-spline}
\usage{
f_life_exp_from_5year_netsurv(
  survival,
  include_additive_hzds,
  include_cure_fraction,
  plotsurvhz,
  survextrap_add_knotts_interval,
  survextrap_fit_method,
  survextrap_niter
)
}
\arguments{
\item{survival}{Data frame input}

\item{include_additive_hzds}{True or False to include background hazard and use additive hazards framework}

\item{include_cure_fraction}{True or False to include cure fraction and use mixture-cure model}

\item{plotsurvhz}{True or False to output plots}

\item{survextrap_add_knotts_interval}{Knots arguement to feed through to survextrap function}

\item{survextrap_fit_method}{Fit method arguement to feed through to survextrap function}

\item{survextrap_niter}{number of iterations for MCMC to feed through to survextrap function}
}
\value{
RMST
}
\description{
Modal life expectancy using M-spline
}
\examples{
f_life_exp_from_5year_netsurv()
}
