% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_prob_death_60.R
\name{f_prob_death_60}
\alias{f_prob_death_60}
\title{Modal life expectancy using M-spline and calc survival}
\usage{
f_prob_death_60(
  survival,
  include_additive_hzds,
  include_cure_fraction,
  plotsurvhz,
  survextrap_add_knotts_interval,
  survextrap_fit_method,
  survextrap_niter
)
}
\arguments{
\item{survival}{Data frame input}

\item{include_additive_hzds}{True or False to include background hazard and use additive hazards framework}

\item{include_cure_fraction}{True or False to include cure fraction and use mixture-cure model}

\item{plotsurvhz}{True or False to output plots}

\item{survextrap_add_knotts_interval}{Knots arguement to feed through to survextrap function}

\item{survextrap_fit_method}{Fit method arguement to feed through to survextrap function}

\item{survextrap_niter}{number of iterations for MCMC to feed through to survextrap function}
}
\value{
Proabability of survival to age 60
}
\description{
Modal life expectancy using M-spline and calc survival
}
\examples{
f_prob_death_60
}
