% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{f_rbeta}
\alias{f_rbeta}
\title{Random sample from beta distribution}
\usage{
f_rbeta(n, p1, p2)
}
\arguments{
\item{n}{number of samples}

\item{p1}{mean}

\item{p2}{sample size or SD/SE}
}
\value{
rounded numeric vector
}
\description{
Random draws from a beta distribution using mean (p1) and either SD/SE or sample size (p2) as arguments
}
\examples{
f_rbeta()
}
