% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{f_rbeta_constrained}
\alias{f_rbeta_constrained}
\title{Random sample from beta distribution constrained by values in samples_high}
\usage{
f_rbeta_constrained(n, p, ss, p_high, ss_high, samples_high)
}
\arguments{
\item{n}{x}

\item{p}{x}

\item{ss}{x}

\item{p_high}{x}

\item{ss_high}{x}

\item{samples_high}{x}
}
\value{
rounded numeric vector
}
\description{
Random draws from a beta distribution constrained by values in samples_high
}
\examples{
f_rbeta_constrained()
}
