% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{f_rgamma}
\alias{f_rgamma}
\title{Random sample from gamma distribution}
\usage{
f_rgamma(n, m, s)
}
\arguments{
\item{n}{number of samples}

\item{m}{mean}

\item{s}{SD/SE}
}
\value{
rounded numeric vector
}
\description{
Random draws from a beta distribution using using mean and SD/SE as arguments
}
\examples{
f_rgamma()
}
