% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_run_model.R
\name{f_run_model}
\alias{f_run_model}
\title{Runs the model using stated inputs}
\usage{
f_run_model(
  no_NICE_modifier = TRUE,
  run_desc = "test",
  export_path = "~/GalleriCEmodel/analyses/NICE_modifier/export/",
  data_dir = "~/GalleriCEmodel/data-raw/England_pilot_planning/",
  return_M_trace = TRUE,
  model_start_age = 50,
  soc_screen_min_age = 50,
  soc_screen_max_age = 85,
  mced_screen_min_age = 50,
  mced_screen_max_age = 79,
  time_horizon = 101,
  perc_female = 0.51,
  disc_health = 0.035,
  disc_cost = 0.035,
  cost_growth_rate_screening = 0,
  cost_growth_rate = 0,
  dur_iatrogenic = 0.25,
  perspective = 1,
  analysis_type = 1,
  cohort_size = 1e+05,
  booAsympStageIII = TRUE,
  booAsympStageIV = TRUE,
  booIatrogenic = 0,
  mced_cost = 949,
  adherence = 1,
  annual_compliance = 0.9,
  mced_specificity = 0.995,
  util_iatrogenic = 0,
  disutil_FP = 0.05,
  disutil_misdiagnosis = 0.05,
  perc_overdiagnosis = 0.05,
  perc_misdiagnosis = 1,
  dur_workup_misdiagnosis = 0.5,
  dur_workup_FP = 0.5,
  iPop = 1,
  start_age_band = 11,
  max_age_band = 20,
  age_band_interval = 5,
  age_option = 1,
  run_type = "deterministic",
  nPSA = 1000,
  nDSA = 18,
  optPSA_soc_test_costs = 1,
  optPSA_costs_cancer_tx = 1,
  optPSA_costs_workup_FP = 2,
  optPSA_costs_workup_misdiagnosis = 2,
  optPSA_util_cancer = 2,
  pctPSA_soc_test_cost = 0,
  pctPSA_costs_cancer_tx = 0,
  pctPSA_costs_workup_FP = 0.2,
  pctPSA_costs_workup_misdiagnosis = 0.2,
  pctPSA_util_cancer = 0.2,
  UtilMultiplier = 1
)
}
\arguments{
\item{run_desc}{Use to name directory in which results will be written. Use test which will be rewritten or specify date}

\item{export_path}{the directory to which results are written}

\item{data_dir}{specify directory in which the data is in}

\item{return_M_trace}{return the markov trace once function is run}

\item{model_start_age}{youngest age patients may start screening}

\item{soc_screen_min_age}{Minimum age of SoC screening start}

\item{soc_screen_max_age}{Maximum SoC screening age}

\item{mced_screen_min_age}{Minimum age of MCED screening start}

\item{mced_screen_max_age}{Maximum MCED screening age}

\item{time_horizon}{Years the model runs over}

\item{perc_female}{= percentage of the population that is female}

\item{disc_health}{= Discount rate applied to health}

\item{disc_cost}{= Discount rate applied to costs}

\item{cost_growth_rate_screening}{Growth rate in the cost of screening test}

\item{cost_growth_rate}{Growth rate in general costs}

\item{dur_iatrogenic}{Duration of survival for patients affected by iatrogenic harm}

\item{perspective}{?? 1}

\item{analysis_type}{1 = general population, 2 = patient heterogeneity}

\item{cohort_size}{size of cohort for markov trace}

\item{booAsympStageIII}{allow stage shift in stage III for patients with asymptomatic cancer}

\item{booAsympStageIV}{allow stage shift in stage IV for patients with asymptomatic cancer}

\item{booIatrogenic}{whether to include iatrogenic harm}

\item{mced_cost}{the cost of the MCED test}

\item{adherence}{adherence to screening}

\item{annual_compliance}{compliance to annual screening}

\item{mced_specificity}{MCED test specificity}

\item{disutil_FP}{disutility associated with false positive diagnosis}

\item{disutil_misdiagnosis}{disutility associated with misdiagnosis}

\item{perc_overdiagnosis}{\% of dead patients with asymptomatic cancer}

\item{perc_misdiagnosis}{\% of patients correctly diagnosed after misdiagnosis}

\item{dur_workup_misdiagnosis}{additional work-up period for misdiagnosis}

\item{dur_workup_FP}{additional work-up period for false positive diagnosis}

\item{iPop}{??VBA index for cohort}

\item{start_age_band}{??VBA index for age}

\item{max_age_band}{??VBA index for age}

\item{age_band_interval}{age band interval size}

\item{age_option}{??1=min age}

\item{run_type}{Either "deterministic", deterministic/probabilistic sensitivity analysis "dsa" or"psa"}

\item{nPSA}{number of simulations for probabilistic sensitivity analysis}

\item{nDSA}{number of parameters for deterministic sensitivity analysis}

\item{optPSA_soc_test_costs}{SoC Costs SE variation option}

\item{optPSA_costs_cancer_tx}{sancer treatment costs variation option}

\item{optPSA_costs_workup_FP}{FP workup costs variation option}

\item{optPSA_costs_workup_misdiagnosis}{misdiagnosis workup costs variation option}

\item{optPSA_util_cancer}{disutility variation option}

\item{pctPSA_soc_test_cost}{value of SoC Costs SE variation option}

\item{pctPSA_costs_cancer_tx}{value of Cancer treatment costs variation option}

\item{pctPSA_costs_workup_FP}{value of FP workup costs variation option}

\item{pctPSA_costs_workup_misdiagnosis}{value of Misdiagnosis workup costs variation option}

\item{pctPSA_util_cancer}{value of Disutility variation option}

\item{UtilMultiplier}{multiplier applied to difference between cancer utility and general population utility}

\item{util_iatrogeni}{the utlity associated with iatrogenic harm}
}
\value{
None - model exports results to "~/GalleriCEmodel/man/..."
}
\description{
Runs the model using stated inputs
}
\examples{
f_run_model()
}
