% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_interception_model_functions.R
\name{rate_given_exptimedist_dwell_screeninginterval}
\alias{rate_given_exptimedist_dwell_screeninginterval}
\title{Calculates event rate for interval}
\usage{
rate_given_exptimedist_dwell_screeninginterval(
  MC_shape_dwell_time_dist,
  event_rate_dwell_interval
)
}
\arguments{
\item{MC_shape_dwell_time_dist}{Monte Carlo generation of the shape of the distribution of the dwell time, based on exp dist}

\item{event_rate_dwell_interval}{Screening interval divided by the dwell time}
}
\value{
Scalar, rate of screening event for a given length of time
}
\description{
Estimates the rate when a screening event happens within randomly distributed length of time
}
\examples{
rate_given_exptimedist_dwell_screeninginterval()
}
