% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_evidera_helper_functions.R
\name{rnd}
\alias{rnd}
\title{Custom round function}
\usage{
rnd(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
rounded numeric vector
}
\description{
Custom round function
}
\examples{
rnd()
}
